update linkis 1.1.1 to 1.3.2、dss 1.1.0 to 1.1.1

+++ server/src/main/java/edp/davinci/dto/dashboardDto/DashboardCreate.java
@@ -37,8 +37,8 @@ public class DashboardCreate {
     @Min(value = 1L, message = "Invalid dashboard portal id")
     private Long dashboardPortalId;
 
-    @Min(value = (short) 0, message = "Invalid dashboard type")
-    @Max(value = (short) 1, message = "Invalid dashboard type")
+    @Min(value = 0, message = "Invalid dashboard type")
+    @Max(value = 1, message = "Invalid dashboard type")
     private Short type = 0;
 
     private Integer index = 0;
